add some cool page

+++ FastdfsMonitor/Controllers/FileController.cs
@@ -1,17 +1,22 @@
-﻿using FastdfsMonitor.Entities;
+﻿using FastDFSCore;
+using FastdfsMonitor.Dtos;
+using FastdfsMonitor.Entities;
 using FastdfsMonitor.Extensions;
+using FastdfsMonitor.Models;
 using FastdfsMonitor.Services;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.VisualBasic.FileIO;
+using Microsoft.EntityFrameworkCore;
+using System.IO;
 
 namespace FastdfsMonitor.Controllers
 {
     public class FileController : Controller
     {
-        private readonly FastDFSService _fastDFSService;
+        private readonly IFastDFSService _fastDFSService;
         private readonly IWebHostEnvironment _webHostEnvironment;
         private readonly RocketDbContext _rocketDbContext;
-        public FileController(FastDFSService fastDFSService, RocketDbContext rocketDbContext,IWebHostEnvironment webHostEnvironment)
+
+        public FileController(IFastDFSService fastDFSService, RocketDbContext rocketDbContext, IWebHostEnvironment webHostEnvironment)
         {
             _fastDFSService = fastDFSService;
             _webHostEnvironment = webHostEnvironment;
@@ -23,15 +28,53 @@ namespace FastdfsMonitor.Controllers
             return View();
         }
 
+        /// <summary>
+        /// 获取数据库文件记录
+        /// </summary>
+        /// <param name="fileId"></param>
+        /// <returns></returns>
+        [HttpPost]
+        public async Task<PageDataResult<FastdfsFileInfo>> GetFilesAsync([FromBody] FastdfsFileQueryDto queryDto)
+        {
+            try
+            {
+                int offsetNum = 0, pagesize = 100;
+                if (queryDto.Page >= 1 && queryDto.Limit > 0)
+                {
+                    offsetNum = (queryDto.Page - 1) * queryDto.Limit;
+                }
+                else
+                {
+                    queryDto.Limit = pagesize;
+                }
+
+                List<FastdfsFileInfo> list = _rocketDbContext.FastDFSFileInfo.OrderBy(b => b.UploadTime).Skip(offsetNum).Take(queryDto.Limit).ToList();
+
+                return new PageDataResult<FastdfsFileInfo>()
+                {
+                    Msg = "success",
+                    Code = 0,
+                    Count = _rocketDbContext.FastDFSFileInfo.Count(),
+                    Data = list
+                };
+            }
+            catch (Exception ex)
+            {
+                return null;
+            }
+        }
+
         [HttpPost]
         public async Task<IActionResult> UploadAsync(IFormFile file)
         {
             try
             {
                 var filePath = Path.Combine(_webHostEnvironment.WebRootPath, "temp", file.FileName);
-                using var fileStream = new FileStream(filePath, FileMode.Create);
-                file.CopyTo(fileStream);
-                var fileId= await _fastDFSService.UploadAsync("group1", filePath);
+                using (var fileStream = new FileStream(filePath, FileMode.Create))
+                {
+                    file.CopyTo(fileStream);
+                }
+                var fileId = await _fastDFSService.UploadAsync("group1", filePath);
 
                 if (!string.IsNullOrEmpty(fileId))
                 {
@@ -49,14 +92,14 @@ namespace FastdfsMonitor.Controllers
                     _rocketDbContext.FastDFSFileInfo.Add(fileInfo);
                     await _rocketDbContext.SaveChangesAsync();
                     System.IO.File.Delete(filePath);
-                    return Ok();
+
+                    return Json(new { Code = 200, Msg = fileId });
+                    //return Ok(fileId);
                 }
                 else
                 {
                     return BadRequest("上传失败");
                 }
-
-               
             }
             catch (Exception ex)
             {
@@ -65,16 +108,25 @@ namespace FastdfsMonitor.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> DownloadAsync(string fileId)
+        public async Task<IActionResult> DownloadAsync(int id = 0)
         {
             try
             {
-                var fileBytes =await _fastDFSService.DownloadAsync("group1", fileId);
-                var fileName = fileId + ".jpg"; // 保存文件的名称
-             
-                if (fileBytes!=null)
+                var fastdfsFileInfo = await _rocketDbContext.FastDFSFileInfo.FirstOrDefaultAsync(x => x.Id == id);
+
+                if (fastdfsFileInfo != null)
                 {
-                    return File(fileBytes, "application/octet-stream", fileName);
+                    var fileId = fastdfsFileInfo.FileId;
+                    var fileBytes = await _fastDFSService.DownloadAsync("group1", fileId);
+
+                    if (fileBytes != null)
+                    {
+                        return File(fileBytes, "application/octet-stream", fastdfsFileInfo.FileName);
+                    }
+                    else
+                    {
+                        return StatusCode(404, "文件不存在");
+                    }
                 }
                 else
                 {
@@ -88,18 +140,55 @@ namespace FastdfsMonitor.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> DeleteAsync(string fileId)
+        public async Task<IActionResult> DeleteAsync(int id)
         {
             try
             {
-                await _fastDFSService.DeleteAsync("group1", fileId);
-                return Ok();
+                var fastdfsFileInfo = await _rocketDbContext.FastDFSFileInfo.FirstOrDefaultAsync(x => x.Id == id); 
+                if (fastdfsFileInfo != null)
+                {
+                    var fileId = fastdfsFileInfo.FileId;
+
+                    #region 开始事务
+                    using (var transaction = _rocketDbContext.Database.BeginTransaction())
+                    {
+                        try
+                        {
+                            // 删除数据库记录
+                            _rocketDbContext.FastDFSFileInfo.Remove(fastdfsFileInfo);
+
+                            // 删除FastDFS文件
+                            var deletedflag = await _fastDFSService.DeleteAsync("group1", fileId);
+                            if (deletedflag)
+                            {
+                                // 提交事务
+                                await _rocketDbContext.SaveChangesAsync();
+                                transaction.Commit();
+                                return Json(new { Code = 200, Msg = "删除图片成功" });
+                            }
+                            else
+                            {
+                                transaction.Rollback();
+                                return Json(new { Code = 500, Msg = "删除图片失败" });
+                            }
+                        }
+                        catch (Exception ex)
+                        {
+                            transaction.Rollback();
+                            return Json(new { Code = 500, Msg = ex.Message });
+                        }
+                    }
+                    #endregion
+                }
+                else
+                {
+                    return Json(new { Code = 404, Msg = "文件不存在" });
+                }
             }
             catch (Exception ex)
             {
-                return StatusCode(500, ex.Message);
+                return Json(new { Code = 500, Msg = ex.Message });
             }
         }
-
     }
 }
